drm/amdgpu: fixed 32-bit kernel build

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_cs.c
@@ -1092,7 +1092,7 @@ int amdgpu_cs_wait_fences_ioctl(struct drm_device *dev, void *data,
 	if (fences == NULL)
 		return -ENOMEM;
 
-	fences_user = (void __user *)wait->in.fences;
+	fences_user = (void __user *)(unsigned long)(wait->in.fences);
 	if (copy_from_user(fences, fences_user,
 		sizeof(struct drm_amdgpu_fence) * fence_count)) {
 		r = -EFAULT;
